ClauseAI_groq/ClauseAI_fixed/backend/requirements.txt
Flask
flask-cors
google-generativeai
python-dotenv
PyPDF2
python-docx
